src/ooga/model/gameBuildingBlocks/validMoveChecks/EmptyColumnCheck.java
package ooga.model.gameBuildingBlocks.validMoveChecks;

import ooga.model.BoardStructure;

/**
 * ValidMoveCheck building block that checks to see if the piece move
 * made is on an empty column.
 */
public class EmptyColumnCheck implements ValidMoveCheck {

  /**
   * Determines if the move is valid
   *
   * @param currBoard BoardStructure to check validity of move for
   * @param pieceI I coordinate of piece to be moved
   * @param pieceJ K coordinate of piece to be moved
   * @param destinationI I coordinate for move destination
   * @param destinationJ J coordinate for move destination
   * @return boolean if the given move passes this valid check
   */
  @Override
  public boolean isValidMove(BoardStructure currBoard, int pieceI, int pieceJ,
      int destinationI, int destinationJ) {
    for (int i = currBoard.getRows() - 1; i >= 0; i--) {
      if (currBoard.getPieceType(i, destinationJ).isEmpty()) {
        return true;
      }
    }
    return false;
  }
}
